Factor out ConceptBuilder
